// recipient.type.ts
export type Recipient  = {
  email_address: string;
  country: string;
  language?: string;
  ip_address?: string;
  subscription_source?: string;
  subscription_date?: string;
  gender?: string;
  title?: string;
  first_name?: string;
  last_name?: string;
  birthdate?: string;
  address?: string;
  zipcode?: string;
  phone1?: string;
  phone2?: string;
  last_activity_date?: string;
  subscription_status?: string;
  unsubscription_ip?: string;
  unsubscription_date?: string;
  complaint?: string;
  complaint_date?: string;
  bounce?: string;
  bounce_date?: string;
  category_ids?: string;
  external_id?: string;
}